TextContentTagElement replaces TextTagElement
*THIS IS BROKEN*

Meant as a conversation opener; amazingly it builds, runs, and works more than I expected.

- Delete TextTagElement.java
- Rename TextContentTagElement as TextTagElement
- Delete ParagraphTagElement; in DefaultTagProvider have <p> really be <text>
- Keep ComponentTagElement, mainly for backwards compatibility; it's class simplifies.
- HeadingTagElement largely stasy the same, as it's render() and getHeight() account for it's use of scaled graphics context
