+++ common/src/main/java/earth/terrarium/hermes/api/defaults/HeadingTagElement.java
@@ -1,11 +1,11 @@
 package earth.terrarium.hermes.api.defaults;
 
 import com.teamresourceful.resourcefullib.client.CloseablePoseStack;
+import com.teamresourceful.resourcefullib.common.color.Color;
 import earth.terrarium.hermes.api.Alignment;
 import earth.terrarium.hermes.api.themes.Theme;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.network.chat.Component;
 import net.minecraft.util.FormattedCharSequence;
 
 import java.util.Arrays;
@@ -27,9 +27,8 @@ public abstract class HeadingTagElement extends TextTagElement {
             float translationFactor = (float) (scale - 1) / scale;
             graphics.pose().translate(-x * translationFactor, -y * translationFactor, 0);
 
-            Component text = Component.nullToEmpty(this.content).copy().setStyle(this.getStyle());
             var font = Minecraft.getInstance().font;
-            var lines = font.split(text, (width - (10 + (2 * xSurround))) / scale);
+            var lines = font.split(component, (width - (10 + (2 * xSurround))) / scale);
 
             var contentWidth = lines.stream().mapToInt((line) -> font.width(line) - 1).max().orElse(0);
             // from top of top row capitals, to bottom of bottom row letters with descenders (eg: "y")
@@ -37,7 +36,7 @@ public abstract class HeadingTagElement extends TextTagElement {
             int[] lineOffsets = lines.stream().mapToInt((line) -> getOffsetForTextTag(width, line)).toArray();
             int contentOffset = Arrays.stream(lineOffsets).min().orElse(width);
 
-            drawFillAndBorder(graphics, x + xSurround + contentOffset, y + ySurround, contentWidth, contentHeight);;
+            drawFillAndBorder(graphics, x + xSurround + contentOffset, y + ySurround, contentWidth, contentHeight);
 
             int lineHeight = font.lineHeight;
             for (int i = 0; i < lines.size(); i++) {
@@ -46,8 +45,8 @@ public abstract class HeadingTagElement extends TextTagElement {
                         lines.get(i),
                         x + xSurround + lineOffsets[i],
                         y + ySurround + (i * (lineHeight + 1)),
-                        this.color,
-                        this.shadowed
+                        Color.DEFAULT,
+                        shadowed
                 );
             }
         }
@@ -55,7 +54,7 @@ public abstract class HeadingTagElement extends TextTagElement {
 
     @Override
     public int getHeight(int width) {
-        int lines = Minecraft.getInstance().font.split(Component.nullToEmpty(this.content), (width - (10 + (2 * xSurround))) / scale).size();
+        int lines = Minecraft.getInstance().font.split(component, (width - (10 + (2 * xSurround))) / scale).size();
         int lineHeight = Minecraft.getInstance().font.lineHeight;
         // scale * (element height + vertical spacing)
         return scale * (((lines * lineHeight) + (lines - 2)) + (2 * ySurround));
